feat: 확정된 급여 조회 기능 개발 (#37) - 자동 송금이 이미 진행됐으면 수정 불가 로직 추가

+++ app/src/main/java/com/mangoboss/app/domain/service/payroll/PayrollService.java
@@ -1,6 +1,8 @@
 package com.mangoboss.app.domain.service.payroll;
 
 
+import com.mangoboss.app.common.exception.CustomErrorInfo;
+import com.mangoboss.app.common.exception.CustomException;
 import com.mangoboss.app.domain.repository.EstimatedPayrollRepository;
 import com.mangoboss.app.domain.repository.PayrollRepository;
 import com.mangoboss.storage.attendance.AttendanceEntity;
@@ -134,6 +136,13 @@ public class PayrollService {
 
     @Transactional
     public void deletePayrollsByStoreIdAndMonth(final Long storeId, final LocalDate month){
+        if(payrollRepository.isNotTransferPending(storeId, month)){
+            throw new CustomException(CustomErrorInfo.PAYROLL_TRANSFER_ALREADY_STARTED);
+        }
         payrollRepository.deleteAllByStoreIdAndMonth(storeId, month);
+    }
+
+    public List<PayrollEntity> getConfirmedPayroll(final Long storeId, final LocalDate month) {
+        return payrollRepository.getAllByStoreIdAndMonth(storeId, month);
     }
 }
